Estimation of Conditionals of Mediators under Exposed

Estimation of Conditionals of Mediators under Exposed

Usage

con_exposed(
  MM,
  k,
  K,
  data,
  dat2,
  fam_type,
  mediators,
  interactions_XC,
  lnzero,
  separation_method,
  n
)

Arguments

MM

An integer specifying the index of the mediator whose distribution will be shifted.

k

An integer specifying the index of the current mediator being processed.

K

An integer specifying the total number of mediators and intermediate confounders.

data

A data.frame containing the dataset for analysis. The dataset is used to fit the model for the mediators.

dat2

A data.frame used to perform counterfactual predictions and random draws.

fam_type

A character string specifying the family type for modeling. Options typically include "gaussian" for continuous variables or "binomial" for binary variables.

mediators

A character vector including the variable names for mediators (including intermediate confounders).

interactions_XC

A character string specifying the two-way interactions amongst exposure and baseline confounders to include in the regression models in the estimation procedure.

lnzero

A numeric vector specifying the non-zero levels of the exposure.

separation_method

Method to handle separation, only relevant for binomial (binary outcome) models. Options are "brglm" (Logistic regression models are fitted using bias reduction methods for generalised linear models implemented in the brglm2 package) or "discard" (if separation is detected, the function returns NA. If this occurs during the main estimation, the program stops; if it occurs during bootstrapping, the affected bootstrap samples are discarded).

n

An integer specifying the number of observations for dat2.