Generate Model Formula for Mediator Models

This function generates a model formula for mediator models based on the input parameters.

Usage

gen_formula(
  k,
  first = NULL,
  MM = NULL,
  K = NULL,
  interactions_XC,
  include_all = FALSE,
  marginal = FALSE
)

Arguments

k

An integer specifying the index of the mediator for which the formula is being generated.

first

An integer (optional) specifying the index of the first mediator. Defaults to NULL, indicating that the function will generate formulas without explicitly considering this parameter.

MM

An integer (optional) specifying the index of the mediator whose distribution will be shifted. Defaults to NULL, indicating that the function will generate formulas without explicitly considering this parameter.

K

An integer (optional) specifying the total number of mediators and intermediate confounders. Defaults to NULL, indicating that the function will generate formulas without explicitly considering this parameter.

interactions_XC

A character string specifying the exposure-confounder or confounder-confounder interaction terms to include in the regression models for confounder adjustment.

include_all

Logical.

marginal

Logical. If TRUE, estimating marginals under X=0.