Estimation of Marginal Distributions of Mediators under Unexposed

Estimation of Marginal Distributions of Mediators under Unexposed

Usage

marg_dist(k, first, K, data, dat2, fam_type, mediators, interactions_XC, n)

Arguments

k

An integer specifying the index of the current mediator being processed.

first

An integer specifying the index of the first mediator.

K

An integer specifying the total number of mediators and intermediate confounders.

data

A data.frame containing the dataset for analysis. The dataset is used to fit the model for the mediators.

dat2

A data.frame used to perform counterfactual predictions and random draws.

fam_type

A character string specifying the family type for modeling. Options typically include "gaussian" for continuous variables or "binomial" for binary variables.

mediators

A character vector including the variable names for mediators (including intermediate confounders).

interactions_XC

A character string specifying the exposure-confounder or confounder-confounder interaction terms to include in the regression models for confounder adjustment.

n

An integer specifying the number of observations for dat2.