Causal mediation analysis for estimating the interventional effect

'medRCT' is used to estimate interventional effects by mapping observational data to a hypothetical target trial. It can handle multiple mediators, including some not of primary interest but that are exposure-induced mediator-outcome confounders.

Usage

medRCT(
  dat,
  exposure,
  outcome,
  mediators,
  intermediate_confs,
  confounders,
  interactions_XC = "all",
  intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
  mcsim,
  bootstrap = TRUE,
  boot_args = list(R = 100, stype = "i", ci.type = "norm"),
  ...
)

Arguments

dat

A data.frame containing the dataset for analysis. It should include variables for the exposure, outcome, mediators, confounders, and exposure-induced mediator-outcome confounders specified in the analysis.

exposure

A character string specifying the name of the exposure variable in the dataset. The exposure variable can be categorical, with 0 explicitly denoting the unexposed (or control) group. Other values represent different exposure categories.

outcome

A character string specifying the name of the outcome variable in the dataset. The outcome variable can be specified as either binary or continuous.

mediators

A character vector specifying the names of mediator variables in the dataset. The mediators can be specified as either binary or continuous. When estimating the effect type "shift_k_order", the order of mediators in the vector is important, as it determines the causal sequence of mediators.

intermediate_confs

A character vector specifying the names of intermediate confounders in the dataset. The intermediate confounders can be specified as either binary or continuous. If NULL, no intermediate confounders are specified, and the natural effect will be estimated.

confounders

A character vector listing the names of baseline confounders.

interactions_XC

A character string specifying the exposure-confounder or confounder-confounder interaction terms to include in the regression models for confounder adjustment. The default value, "all", includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions. Specify "none" to exclude all two-way exposure-confounder and confounder-confounder interactions.

intervention_type

A character string indicating the type of interventional effect to be estimated. Options include:

"all" (default): Estimates all types of interventional indirect effects.

"shift_all": Estimates the interventional indirect effect of shifting the joint distribution of all mediators in the exposed to match the level in the unexposed.

"shift_k": Estimates the interventional indirect effect of shifting the distribution of a specific mediator (k) in the exposed to match the level in the unexposed.

"shift_k_order": Estimates the interventional indirect effect of shifting the distribution of a specific mediator (k) in the exposed to match the level in the unexposed while accounting for the flow-on effects on its causally descendent mediators.

mcsim

An integer specifying the number of Monte Carlo simulations to perform.

bootstrap

A logical value indicating whether bootstrapping should be performed. If TRUE (default), bootstrapping is conducted using the boot function from the boot package.

boot_args

A list of arguments for bootstrapping. The default settings are:

R: Number of bootstrap replicates (default: 100).

stype: Specifies the statistic type passed to the boot function (default: "i").

ci.type: Specifies the type of confidence interval to compute (default: "norm").

...

Additional arguments passed to the boot function from the boot package.