Causal mediation analysis for estimating the interventional effect mapped to a target trial

Causal mediation analysis for estimating the interventional effect mapped to a target trial

Usage

medRCT.fun(
  dat,
  ind = 1:nrow(dat),
  first = first,
  K = K,
  fam_type = fam_type,
  interactions_XC = interactions_XC,
  intervention_type = intervention_type,
  mcsim
)

Arguments

dat

A data.frame with the data for analysis

ind

A vector of indices that define the sample from dat on which to conduct the analysis. Defaults to all rows of the data. This facilitates the use of this function within the boot() function from the boot package

first

index of first mediator of interest after combining intermediate confounders with mediators

K

the number of mediators

fam_type

family type for mediators

interactions_XC

a character string specifying the exposure-confounder or confounder-confounder interaction terms to include in all regression models in the procedure. Defaults to include all two-way exposure-confounder interactions and no confounder-confounder interactions.

intervention_type

a character string indicating the type of the interventional effect to be estimated. Can be 'all', 'shift_all', 'shift_k', 'shift_k_order'. Default is 'all'.

mcsim

the number of Monte Carlo simulations to conduct