Set Exposure Column to a Specific Value and Convert to Factor

Set Exposure Column to a Specific Value and Convert to Factor

Usage

set_exposure(data, column_name, exp_val)

Arguments

data

A data.table object.

column_name

A character string specifying the name of the column to modify.

exp_val

The exposure value to assign to the column.