Causal mediation analysis for estimating the interventional effect

'medRCT' is used to estimating the interventional effects that are defined by a mapping to a target trial. It can handle multiple mediators, including some not of primary interest but that are intermediate confounders

Usage

medRCT(
  dat,
  exposure,
  outcome,
  mediators,
  intermediate_confs,
  confounders,
  interactions_XC = "all",
  intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
  mcsim,
  bootstrap = TRUE,
  boot_args = list(R = 100, stype = "i", ci.type = "norm"),
  ...
)

Arguments

dat

a data.frame with the data for analysis

exposure

a character string representing the name of the exposure variable in the data. The exposure variable must be binary, with 1 indicating exposed or treated, and 0 indicating unexposed or control.

outcome

a character string representing the name of the outcome variable in the data. The outcome variable must be binary.

mediators

a character vector representing the names of the mediators in the data. The mediators must be binary variables.

intermediate_confs

a character vector representing the names of the intermediate confounders in the data. The intermediate confounders must be binary variables.

confounders

a character vector representing the names of the confounders in the data, which must be of the required class (e.g. factor if appropriate)

interactions_XC

a character string specifying the exposure-confounder or confounder-confounder interaction terms to include in all regression models in the procedure. Default is 'all', which includes all two-way exposure-confounder interactions and excludes confounder-confounder interactions.

intervention_type

a character string indicating the type of the interventional effect to be estimated. Can be 'all', 'shift_all', 'shift_k', 'shift_k_order'. Default is 'all', under which all effects will be estimated. ‘shift all’ is to estimate the interventional effect under shifting the joint distribution of all mediatios. ‘shift k’ is to estimate the interventional effect under shifting the distribution of mediator k, independently of other mediators. ‘shift_k_order’ is to estimate the interventional effect under shifting the distribution of mediator k, independently of the its antecedent mediators, while allowing for the flow on effect of mediator k on its descendent mediators.

mcsim

the number of Monte Carlo simulations to conduct

bootstrap

logical. If TRUE, bootstrap will be conducted.

boot_args

a list of bootstrapping arguments. R is the number of bootstrap replicates. stype indicates what the second argument of statistics in the boot function represents

...

other arguments passed to the boot function in the boot package.