Causal Mediation Analysis for Estimating Interventional Effects

This function performs the actual causal mediation analysis to estimate interventional effects mapped to a hypothetical target trial.

Usage

medRCT.fun(
  dat,
  ind = 1:nrow(dat),
  first = first,
  K = K,
  fam_type = fam_type,
  interactions_XC = interactions_XC,
  intervention_type = intervention_type,
  mcsim
)

Arguments

dat

A data.frame containing the dataset for analysis.

ind

A vector of indices specifying the subset of dat to use for the analysis. Defaults to all rows of dat. This parameter is particularly useful when using this function within the boot() function from the boot package, as it enables resampling by specifying subsets of the data.

first

An integer specifying the index of the first mediator of interest in the combined list of intermediate confounders and mediators.

K

An integer specifying the total number of mediators and intermediate confounders. Mediators are considered sequentially based on their order.

fam_type

A character string specifying the family type for modeling. Options typically include "gaussian" for continuous variables or "binomial" for binary variables.

interactions_XC

A character string specifying the exposure-confounder or confounder-confounder interaction terms to include in the regression models for confounder adjustment. The default value, "all", includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions. Specify "none" to exclude all two-way exposure-confounder and confounder-confounder interactions.

intervention_type

A character string indicating the type of interventional effect to be estimated. Options include:

"all" (default): Estimates all types of interventional indirect effects.

"shift_all": Estimates the interventional indirect effect of shifting the joint distribution of all mediators in the exposed to match the level in the unexposed.

"shift_k": Estimates the interventional indirect effect of shifting the distribution of a specific mediator (k) in the exposed to match the level in the unexposed.

"shift_k_order": Estimates the interventional indirect effect of shifting the distribution of a specific mediator (k) in the exposed to match the level in the unexposed while accounting for the flow-on effects on its causal descendent mediators.

mcsim

An integer specifying the number of Monte Carlo simulations to perform.